Refactors calldata into get tx info
