+++ packages/ethernaut-interact/README.md
@@ -41,7 +41,7 @@ This plugin adds the following tasks:
 - logs Finds logs emitted by a contract
 - send Sends ether to an address
 - token Interacts with any ERC20 token
-- calldata Decodes calldata from a transaction id
+- tx Gives information about a mined transaction
 
 ## Environment extensions
