+++ packages/ethernaut-interact/test/fixture-projects/basic-project/abis/dai.json
